+++ src/interface/IGlobalConfig.ts
@@ -1,4 +1,6 @@
 export interface IGlobalConfig {
 	/** @type {string} The API Key */
 	apiKey?: string;
+	/** @type {string} The API Url */
+	apiUrl?: string;
 }