+++ src/interface/schema/GlobalConfig.ts
@@ -1,7 +1,6 @@
 import * as Joi from 'joi';
 
 export const GlobalConfigSchema = Joi.object({
-	apiKey: Joi.string()
-		.length(48)
-		.description('test')
+	apiKey: Joi.string().length(48),
+	apiUrl: Joi.string().min(1)
 });